requirements.txt
python-telegram-bot==21.6
pytz==2024.1  # для совместимости; основной TZ — через zoneinfo
